shop_product_service/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
typing_extensions
pytest
pytest-django
requests
django-cors-headers
coreapi
coreschema
django-filter
httpie